requirements.txt
PySide6>=6.5
pyqtgraph>=0.13
pyserial>=3.5
keyring>=24.0
